backend/requirements.txt
Django>=5.0
djangorestframework
celery
redis
openai
langchain
psycopg2-binary  # PostgreSQL driver
